+++ public/app/plugins/datasource/mysql/specs/datasource_specs.ts
@@ -88,7 +88,7 @@ describe('MySQLDatasource', function() {
           refId: 'tempvar',
           tables: [
             {
-              columns: [{text: 'time_sec'}, {text: 'title'}, {text: 'text'}],
+              columns: [{text: 'title'}, {text: 'text'}],
               rows: [
                 ['aTitle', 'some text'],
                 ['aTitle2', 'some text2'],
@@ -112,6 +112,86 @@ describe('MySQLDatasource', function() {
       expect(results.length).to.be(6);
       expect(results[0].text).to.be('aTitle');
       expect(results[5].text).to.be('some text3');
+    });
+  });
+
+  describe('When performing metricFindQuery with key, value columns', function() {
+    let results;
+    const query = 'select * from atable';
+    const response = {
+      results: {
+        tempvar: {
+          meta: {
+            rowCount: 3
+          },
+          refId: 'tempvar',
+          tables: [
+            {
+              columns: [{text: '__value'}, {text: '__text'}],
+              rows: [
+                ['value1', 'aTitle'],
+                ['value2', 'aTitle2'],
+                ['value3', 'aTitle3']
+              ]
+            }
+          ]
+        }
+      }
+    };
+
+    beforeEach(function() {
+      ctx.backendSrv.datasourceRequest = function(options) {
+        return ctx.$q.when({data: response, status: 200});
+      };
+      ctx.ds.metricFindQuery(query).then(function(data) { results = data; });
+      ctx.$rootScope.$apply();
+    });
+
+    it('should return list of as text, value', function() {
+      expect(results.length).to.be(3);
+      expect(results[0].text).to.be('aTitle');
+      expect(results[0].value).to.be('value1');
+      expect(results[2].text).to.be('aTitle3');
+      expect(results[2].value).to.be('value3');
+    });
+  });
+
+  describe('When performing metricFindQuery with key, value columns and with duplicate keys', function() {
+    let results;
+    const query = 'select * from atable';
+    const response = {
+      results: {
+        tempvar: {
+          meta: {
+            rowCount: 3
+          },
+          refId: 'tempvar',
+          tables: [
+            {
+              columns: [{text: '__text'}, {text: '__value'}],
+              rows: [
+                ['aTitle', 'same'],
+                ['aTitle', 'same'],
+                ['aTitle', 'diff']
+              ]
+            }
+          ]
+        }
+      }
+    };
+
+    beforeEach(function() {
+      ctx.backendSrv.datasourceRequest = function(options) {
+        return ctx.$q.when({data: response, status: 200});
+      };
+      ctx.ds.metricFindQuery(query).then(function(data) { results = data; });
+      ctx.$rootScope.$apply();
+    });
+
+    it('should return list of unique keys', function() {
+      expect(results.length).to.be(1);
+      expect(results[0].text).to.be('aTitle');
+      expect(results[0].value).to.be('same');
     });
   });
 });
